Enable public url-addressing (#1)
* Set the uuid in before_create

This way we only make one database update

* Refactor database query into a concern

to re-use the ExportsController, as well as the many query page.

* Link to the CSV version of the ExportsController

* isolate "included" to only reference helper_method :blazer_time_value
